watterson: run the Doppler filter in double, it diverges at high sample rates
The 2nd-order Butterworth that shapes each path's Doppler tap kept its state and
coefficients in float. At a high sample rate with a slow Doppler the cutoff
ratio fc/fs becomes tiny and the poles sit within ~1e-4 of the unit circle,
where float's ~6e-8 of relative resolution is enough to push a pole OUTSIDE it.
The filter then grows without bound instead of fading, and the "faded signal"
reported is the blow-up.

Measured before the change, two paths at 1 Hz: 8 kHz gave a sane SNR3k of
-8.37 dB, 48 kHz gave +65.76 dB.

Simulation-only -- nothing on the air uses this -- but it silently corrupts any
fading measurement taken above 8 kHz, which is where mode comparisons and
threshold decisions come from. Mercury's own bench runs at 8 kHz and is
unaffected; the risk is to anyone who reaches for a higher rate and believes the
result.

tests/common/test_watterson.c pins two properties across 8/16/24/48/96 kHz and
0.1-2 Hz: the faded output stays finite, and the measured SNR3k lands in a
plausible band rather than diverging. A single static path is additionally
required to be rate-invariant to 0.5 dB.

Verified to fail against the float version -- but only once the mutation was
complete. A partial revert does not compile (iir_tick's callers and the impulse
-response accumulator take the same types), and a failed build silently leaves
the previous binary in place, which looks exactly like a test that does not
catch the regression. It does.

+++ common/watterson.h
@@ -44,14 +44,21 @@ typedef struct watterson_path
 
     /* IIR filter state for Doppler shaping (2nd-order Butterworth LPF)
      * Separate filters are used for the I and Q components */
-    float x_i[3];        /* Input history, I component */
-    float y_i[3];        /* Output history, I component */
-    float x_q[3];        /* Input history, Q component */
-    float y_q[3];        /* Output history, Q component */
+    /* double, not float: at a high sample rate with a slow Doppler the
+     * cutoff ratio fc/fs becomes tiny and the Butterworth poles sit within
+     * ~1e-4 of the unit circle. float carries only ~6e-8 of relative
+     * resolution there, which is enough to push a pole OUTSIDE the circle:
+     * the filter then diverges instead of fading. Measured before the change,
+     * 2 paths at 1 Hz: 8 kHz gave a sane SNR3k of -8.37 dB but 48 kHz gave
+     * +65.76 dB, i.e. the "signal" was the filter blowing up. */
+    double x_i[3];       /* Input history, I component */
+    double y_i[3];       /* Output history, I component */
+    double x_q[3];       /* Input history, Q component */
+    double y_q[3];       /* Output history, Q component */
 
     /* IIR filter coefficients (a0 is normalised to 1.0) */
-    float b0, b1, b2;    /* Numerator */
-    float a1, a2;        /* Denominator (a0 = 1) */
+    double b0, b1, b2;   /* Numerator   (double: see the history arrays) */
+    double a1, a2;       /* Denominator (a0 = 1) */
 
     /* Frequency offset state */
     float phase;          /* Current phase accumulator for rotation */
